+++ program/src/ids.rs
@@ -27,3 +27,8 @@ pub mod mngo_token {
     #[cfg(not(any(feature = "devnet", feature = "testnet")))]
     declare_id!("MangoCzJ36AjZyKwVj3VnYU4GTonjfVEnJmvvWaxLac");
 }
+
+pub mod luna_pyth_oracle {
+    use solana_program::declare_id;
+    declare_id!("5bmWuR1dgP4avtGYMNKLuxumZTVKGgoN2BCMXWDNL9nY");
+}